improve authorisation in a better way

+++ app/Http/Controllers/ManagerController.php
@@ -15,9 +15,7 @@ use Illuminate\Support\Facades\Gate;
 class ManagerController extends Controller
 {
     public function index() {
-        if (! Gate::allows('visit-manage-computers-page')) {
-            abort(403);
-        }
+        Gate::authorize('visit-manage-computers-page');
 
         $computers = Computer::all();
 
@@ -63,14 +61,18 @@ class ManagerController extends Controller
     }
 
     public function deleteComputer($toDelPCID) {
+        if (Computer::find($toDelPCID)->renting()) {
+            $computerName = Computer::find($toDelPCID)->name;
+            return back()->with('warningMsg',
+                $computerName.' has been rented out. Please confirm return first then can do the delete.');
+        }
+
         Computer::find($toDelPCID)->delete();
         return json_encode("delete success");
     }
 
     public function staffUserManagement() {
-        if (! Gate::allows('visit-staff-manage-users-page')) {
-            abort(403);
-        }
+        Gate::authorize('visit-staff-manage-users-page');
 
         $blackUserIDs = BlackHistory::selectRaw('user_id')
             ->groupBy('user_id')
@@ -88,9 +90,7 @@ class ManagerController extends Controller
     }
 
     public function adminUserManagement() {
-        if (! Gate::allows('visit-admin-manage-users-page')) {
-            abort(403);
-        }
+        Gate::authorize('visit-admin-manage-users-page');
 
         $blackUserIDs = BlackHistory::selectRaw('user_id')
             ->groupBy('user_id')
@@ -141,9 +141,7 @@ class ManagerController extends Controller
     }
 
     public function adminDashboard() {
-        if (! Gate::allows('visit-admin-dashboard-page')) {
-            abort(403);
-        }
+        Gate::authorize('visit-admin-dashboard-page');
 
         $privilegeStudent = Privilege::select('id')->where('name', 'student')->first();
         $privilegeStaff = Privilege::select('id')->where('name', 'customer')->first();
